Got rid of warnings

+++ Project_IP-3/library/src/Core/Processer.cpp
@@ -7,7 +7,7 @@ int Processer::truncate(double value)
 	if (value < 0) return 0;
 	if (value > 255) return 255;
 
-	return value;
+	return (int)value;
 }
